feat(nvim): update plugins and key mappings
This commit updates various plugins in the lazy-lock.json file and
modifies some key mappings in keymaps.lua. It also introduces a new
plugin configuration for mini.diff in git-diff.lua and comments out
the configuration for precognition.nvim. The cmp-config.lua file is
also updated to adjust some formatting and comments.