feat(README): Refactor README and add installation guides
This commit refactors the README.md file to provide a more structured
and detailed overview of the project. It also adds two new markdown
files, arch-install.md and after-install.md, which provide detailed
guides for installing Arch Linux and setting up the system after
installation. The changes aim to make the project more accessible and
easier to set up for new users.